Add enum and Fallible support
The zoura macro now works on 3 more types:
- c-like byte exhaustive
- c-like bit exhaustive
- c-like unexhaustive

"Exhaustive" means the enum has a variant for each possible value of its
representation. Since being byte-exhaustive means you are also
bit-exhaustive, such enums need not be fallible, while the other kinds
should (either byte fallibility or both).

When an enum is "fallible", it can be wrapped in the Fallible type,
which stores an enum that may possibly have an invalid value.